Introduce `Peer::decapsulate`
